+++ ICD.Connect.Misc.CrestronPro/Devices/Ethernet/CrestronEthernetDeviceAdapterNetworkInfo.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Text.RegularExpressions;
 using ICD.Connect.Devices.Telemetry.DeviceInfo;
 
@@ -6,8 +7,15 @@ namespace ICD.Connect.Misc.CrestronPro.Devices.Ethernet
 {
 	public struct CrestronEthernetDeviceAdapterNetworkInfo : IEquatable<CrestronEthernetDeviceAdapterNetworkInfo>
 	{
+		#region Constants
+
+		private const string ETHERNET_ADAPTER_REGEX = @"Ethernet Adapter \[(?'AdapterName'\S+)\]:\s+(?:Link Status.+:\s+(?'LinkStatus'\S+))\s*(?:DHCP\s+.+:\s+(?'DHCP'\S+))\s*(?:MAC Address(?:\(es\))*\s*.+:\s+(?'MACAddress'\S+))\s*(?:IP Address\s+.+:\s+(?'IPV4'\S+))\s*(?:Subnet Mask\s+.+:\s+(?'SubnetMask'\S+))\s*(?:(?:IPv6|IPV6) Address\s*.+:\s+(?'IPV6'\S+)?)\s*(?:Default Gateway\s+.+:\s+(?'DefaultGateway'\S+))\s*(?:DNS Servers\s*.+:\s+(?'DNS'\S+))";
+
+		#endregion
+
 		#region Fields
 
+		private readonly string m_AdapterName;
 		private readonly bool m_Dhcp;
 		private readonly IcdPhysicalAddress m_MacAddress;
 		private readonly string m_IpAddress;
@@ -18,6 +26,8 @@ namespace ICD.Connect.Misc.CrestronPro.Devices.Ethernet
 		#endregion
 
 		#region Properties
+
+		public string AdapterName {get { return m_AdapterName; }}
 
 		public bool Dhcp { get { return m_Dhcp; } }
 
@@ -38,15 +48,17 @@ namespace ICD.Connect.Misc.CrestronPro.Devices.Ethernet
 		/// <summary>
 		/// Constructor.
 		/// </summary>
+		/// <param name="adapterName"></param>
 		/// <param name="dhcp"></param>
 		/// <param name="macAddress"></param>
 		/// <param name="ipAddress"></param>
 		/// <param name="subnetMask"></param>
 		/// <param name="defaultGateway"></param>
 		/// <param name="dnsServer"></param>
-		public CrestronEthernetDeviceAdapterNetworkInfo(bool dhcp, IcdPhysicalAddress macAddress, string ipAddress, string subnetMask,
+		public CrestronEthernetDeviceAdapterNetworkInfo(string adapterName, bool dhcp, IcdPhysicalAddress macAddress, string ipAddress, string subnetMask,
 		                                                string defaultGateway, string dnsServer)
 		{
+			m_AdapterName = adapterName;
 			m_Dhcp = dhcp;
 			m_MacAddress = macAddress == null ? null : macAddress.Clone();
 			m_IpAddress = ipAddress;
@@ -60,22 +72,32 @@ namespace ICD.Connect.Misc.CrestronPro.Devices.Ethernet
 		/// </summary>
 		/// <param name="match"></param>
 		/// <returns></returns>
-		public static CrestronEthernetDeviceAdapterNetworkInfo Parse(Match match)
+		public static CrestronEthernetDeviceAdapterNetworkInfo[] Parse(Match match)
 		{
 			if (!match.Success)
 				throw new InvalidOperationException("Unable to find a matching pattern in IP Config data");
 
-			bool dhcp = match.Groups["DHCP"].Value.Equals("ON", StringComparison.OrdinalIgnoreCase);
-			string mac = match.Groups["MACAddress"].Value;
-			string ip = match.Groups["IPV4"].Value;
-			string mask = match.Groups["SubnetMask"].Value;
-			string gateway = match.Groups["DefaultGateway"].Value;
-			string dns = match.Groups["DNS"].Value;
+			Regex regex = new Regex(ETHERNET_ADAPTER_REGEX);
+			MatchCollection groups = regex.Matches(match.Value);
 
-			IcdPhysicalAddress macAddress;
-			IcdPhysicalAddress.TryParse(mac, out macAddress);
+			List<CrestronEthernetDeviceAdapterNetworkInfo> adapters = new List<CrestronEthernetDeviceAdapterNetworkInfo>();
+			foreach (Match m in groups)
+			{
+				string name = m.Groups["AdapterName"].Value;
+				bool dhcp = m.Groups["DHCP"].Value.Equals("ON", StringComparison.OrdinalIgnoreCase);
+				string mac = m.Groups["MACAddress"].Value;
+				string ip = m.Groups["IPV4"].Value;
+				string mask = m.Groups["SubnetMask"].Value;
+				string gateway = m.Groups["DefaultGateway"].Value;
+				string dns = m.Groups["DNS"].Value;
 
-			return new CrestronEthernetDeviceAdapterNetworkInfo(dhcp, macAddress, ip, mask, gateway, dns);
+				IcdPhysicalAddress macAddress;
+				IcdPhysicalAddress.TryParse(mac, out macAddress);
+
+				adapters.Add(new CrestronEthernetDeviceAdapterNetworkInfo(name, dhcp, macAddress, ip, mask, gateway, dns));
+			}
+
+			return adapters.ToArray();
 		}
 
 		#endregion